pyproject.toml
[project]
name = "porn-video-downloader"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "attrs==25.1.0",
    "beautifulsoup4==4.13.1",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "h11==0.14.0",
    "idna==3.10",
    "outcome==1.3.0.post0",
    "pyee==12.1.1",
    "pysocks==1.7.1",
    "python-ffmpeg==2.0.12",
    "pyyaml>=6.0.3",
    "requests==2.32.3",
    "selenium==4.28.1",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "tqdm==4.67.1",
    "trio==0.28.0",
    "trio-websocket==0.11.1",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
]

[dependency-groups]
dev = [
    "pyright[nodejs]>=1.1.409",
]
